Transaction: Avoid throwing exceptions unpredictably
Transactions can be canceled at any time. If an exception ends up
getting handled by the transaction after this, then then there
is no where to route the exception to. The previous way to handle
this was to regrow it, but this leads to really unpredictable 
behavior. The new behavior is to simply log it and continue.