OptionSet: Fix bug in getUri() that corrupts the query
The method `OptionSet#getURI()` was corrupting the query component of
the URI. For example, a query component of `d=2` would be corrupted
into `d%3D2`.

The root cause is the overzealous use of `Utils.uriEscapeString()`,
which will escape the character `=` to `%3D`.

This change addresses the issue by converting the first instance of
the string `"%3D"` back to `"="` if the query component originally
contained an equal sign. Other approaches tended to be more difficult
to read and review.

A unit test is included which reproduces the original issue and
demonstrates that these changes address the issue.