RetransmissionLayer: Don't retransmit obsolete messages
The retransmission layer was retransmitting obsolete messages, which
was causing problems with `Observable`: when there was lots of
traffic, sometimes a packet would timeout and cancel the observable.

This change makes it so that sending a request with the same outbound
message handler will cause the previous message with that outbound
message handler to be dropped. In some future revisions to this
library, that idea will be expanded to more easily allow the
retransmission of specific sent packets to be cancelled. But that is
for another day.